requirements.txt
# Base dependencies
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.12.1
cryptography==40.0.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
flask-cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.47.2
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
PyJWT==2.6.0
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
psutil==5.9.5
PyMySQL==1.0.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-keycloak==3.9.0
reportlab==4.0.9
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
WTForms==3.1.2
pytest==7.3.1
pyotp==2.9.0
pytest-mock==3.14.0
gunicorn==22.0.0
eventlet==0.35.2

# Django dependencies
Django>=4.2,<5.0
django-allauth>=0.54.0
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.1.0
django-crispy-forms>=2.0
django-rest-framework>=0.1.0
djangorestframework>=3.14.0

# gRPC dependencies
grpcio>=1.59.0
grpcio-tools>=1.59.0
protobuf>=4.24.0
grpcio-reflection>=1.59.0
grpcio-status>=1.59.0

# Modern frontend dependencies
node>=18.0.0
npm>=9.0.0
react>=18.0.0
react-dom>=18.0.0
react-router-dom>=6.0.0
redux>=4.2.0
axios>=1.5.0
